KA-520 Reorganize GetContentItemModel method
My intention here was not completely refactor the method, just name parameter more accuretly and split the monolitic method into several smaller methods with roughly similar internal level of abstraction.

There is bunch of logic that needs to be thought through and refactored properly regards recursion if we want the class and method to somehow read nicely. I believe there would need to be a set of classes and small framework to process ContentItem values, yet it needs much more thinking through that this simple reorganization. Better test coverage would also be neat – I believe my own changes to to be on about the borderline of changes we can afford with current level of (scenarios) coverage.